requirements.txt
argparse
matplotlib
numpy
pandas
tensorflow_hub
streamlit
opencv-contrib-python
opencv-python
pillow
